+++ issuer-service/src/main/java/ch/admin/bj/swiyu/issuer/service/credential/HolderBindingService.java
@@ -121,8 +121,8 @@ public class HolderBindingService {
         if (batchCredentialIssuanceMetadata == null && proofs.size() > 1) {
             throw new Oid4vcException(INVALID_PROOF, "Multiple proofs are not allowed for this credential request");
         }
-        if (batchCredentialIssuanceMetadata != null && batchCredentialIssuanceMetadata.batchSize() != proofs.size()) {
-            throw new Oid4vcException(INVALID_PROOF, "The number of proofs must match the batch size");
+        if (batchCredentialIssuanceMetadata != null && batchCredentialIssuanceMetadata.batchSize() < proofs.size()) {
+            throw new Oid4vcException(INVALID_PROOF, "The number of proofs must be at least the same as the batch size");
         }
     }